Move new code to new .go file
